requirements.txt
fastapi
uvicorn
openai
python-multipart
python-dotenv
httpx